New email property trait.
