+++ src/Entity/Property/EmailPropertyAwareTrait.php
@@ -1,0 +1,41 @@
+<?php
+
+/*
+ * Copyright 2018 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+namespace Apigee\Edge\Entity\Property;
+
+trait EmailPropertyAwareTrait
+{
+    /** @var string */
+    protected $email;
+
+    /**
+     * @inheritdoc
+     */
+    public function getEmail(): ?string
+    {
+        return $this->email;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function setEmail(string $email): void
+    {
+        $this->email = $email;
+    }
+}
