+++ src/Entity/Property/EmailPropertyInterface.php
@@ -1,0 +1,36 @@
+<?php
+
+/*
+ * Copyright 2018 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+namespace Apigee\Edge\Entity\Property;
+
+interface EmailPropertyInterface
+{
+    /**
+     * Returns the email address.
+     *
+     * @return null|string
+     */
+    public function getEmail(): ?string;
+
+    /**
+     * Sets the email address.
+     *
+     * @param string $email
+     */
+    public function setEmail(string $email): void;
+}
